requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
kafka-python